........
  r10801 | holger | 2008-02-14 12:29:23 +0100 (Do, 14 Feb 2008) | 1 line
  
  typecasts are useless
........
  r10802 | holger | 2008-02-14 12:39:16 +0100 (Do, 14 Feb 2008) | 1 line
  
  mowing down more warnings
........
  r10803 | holger | 2008-02-14 14:31:25 +0100 (Do, 14 Feb 2008) | 1 line
  
  more lawnmowing
........


git-svn-id: https://svn.codehaus.org/mule/branches/mule-2.x@10814 bf997673-6b11-0410-b953-e057580c5b09

+++ transports/tcp/src/test/java/org/mule/transport/tcp/integration/AbstractStreamingCapacityTestCase.java
@@ -106,7 +106,7 @@ public abstract class AbstractStreamingCapacityTestCase extends FunctionalTestCa
         long timeEnd = System.currentTimeMillis();
         double speed = size / (double) (timeEnd - timeStart) * 1000 / ONE_MB;
         logger.info("Transfer speed " + speed + " MB/s (" + size + " B in " + (timeEnd - timeStart) + " ms)");
-        double usePercent = 100.0 * delta / ((double) size);
+        double usePercent = 100.0 * delta / size;
         logger.info("Memory delta " + delta + " B = " + usePercent + "%");
         assertTrue("Memory used too high", usePercent < 10);
